+++ libraries/avatars/src/AvatarData.h
@@ -303,9 +303,9 @@ public slots:
     void loadRecording(QString filename);
     void startPlaying();
     void setPlayerVolume(float volume);
-    void setPlayerAudioOffset(qint64 audioOffset);
-    void setPlayerFrame(int frame);
-    void setPlayerTime(qint64 time);
+    void setPlayerAudioOffset(int audioOffset);
+    void setPlayerFrame(unsigned int frame);
+    void setPlayerTime(unsigned int time);
     void setPlayFromCurrentLocation(bool playFromCurrentLocation);
     void setPlayerLoop(bool loop);
     void setPlayerUseDisplayName(bool useDisplayName);
